iceplant_portal/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
pandas==2.2.3
openpyxl==3.1.5
reportlab==4.3.1
psycopg2-binary==2.9.10
Pillow==11.1.0
python-dateutil==2.9.0.post0
pytz==2025.1
django-filter==24.1 